requirements.txt
fastapi
uvicorn
nltk
numpy
sqlalchemy
pydantic
alembic
psycopg2
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]